### 2020-03-04 #2141 @whedon
[PRE REVIEW]: PyZFS: A Python package for first-principles calculations of zero-field splitting tensors
**Submitting author:** @hema-ted (<a href="http://orcid.org/0000-0001-8987-8562">He Ma</a>)
**Repository:** <a href="https://github.com/hema-ted/pyzfs" target ="_blank">https://github.com/hema-ted/pyzfs</a>
**Version:** v1.0
