### 2020-03-11 #2160 @dfm
@xwang862, @malramsay64: Your reviewer instructions and checklists are above and there's more info on the documentation page: https://joss.readthedocs.io/en/latest/index.html

Please don't hesitate to ask if you have any questions throughout this process!
