### 2020-03-13 0c48667
improve installation instructions
