### 2020-03-13 d0793ec
update doc to reflect the use of entry point suggested by @malramsay64
