+++ examples/diamond_nv_qe_hdf5/run.sh
@@ -5,7 +5,9 @@
 # QE must be compiled with HDF5 flag enabled.
 
 mpirun pw.x -i pw.in > pw.out
-mpirun python -m pyzfs.exec.runzfs --wfcfmt qeh5 > zfs.out
+
+mpirun pyzfs --wfcfmt qeh5 > zfs.out
+# equivalently: mpirun python -m pyzfs.exec.runzfs --wfcfmt qeh5 > zfs.out
 
 D=`grep --color=never "D unit" zfs.xml | grep --color=never -Eoh '[+-]?[0-9]+([.][0-9]+)?'`
 Dref=`grep --color=never "D unit" zfs_ref.xml | grep --color=never -Eoh '[+-]?[0-9]+([.][0-9]+)?'`
