+++ doc/tutorial.rst
@@ -20,21 +20,33 @@ After installation, **PyZFS** can be executed in two manners:
 
    Example Jupyter notebooks can be found at /examples/o2_qbox_xml/run.ipynb and /examples/o2_qe_hdf5/run.ipynb.
 
-2. Directly execute *pyzfs.exec.runzfs* with `python -m`. This approach works more smoothly with MPI. 
+2. Directly execute **PyZFS**. This approach works more smoothly with MPI.
    
    For serial execution, simply type the following command in the folder that contains DFT wavefunction file(s)
 
    .. code:: bash
 
-      $ python -m pyzfs.exec.runzfs [--flags]
+      $ pyzfs [--flags]
 
    For parallel execution, use the following command
 
    .. code:: bash
 
-      $ mpiexec [-n num_of_processes] python -m pyzfs.exec.runzfs [--flags]
+      $ mpiexec [-n num_of_processes] pyzfs [--flags]
 
    where *num_of_processes* is the number of processes. **PyZFS** distributes the calculations on a square grid of processes. If *num_of_processes* is not a square number, **PyZFS** will use the largest square number of processes smaller than *num_of_processes* for calculations.
+
+   Note that to use the above `pyzfs` command, one needs to install **PyZFS** through **pip** (see :ref:`installation`). If one manually added **PyZFS** directory to the **PYTHONPATH** without installing it, one needs to replace the above commands with
+
+   .. code:: bash
+
+      $ python -m pyzfs.exec.runzfs [--flags]
+
+   and
+
+   .. code:: bash
+
+      $ mpiexec [-n num_of_processes] python -m pyzfs.exec.runzfs [--flags]
 
    Acceptable flags [--flags] are listed below, for detailed explanation see *pyzfs/exec/runzfs.py*.
 
@@ -58,7 +70,7 @@ After installation, **PyZFS** can be executed in two manners:
 
    .. code:: bash
 
-      $ mpiexec python -m pyzfs.exec.runzfs --wfcfmt qeh5 --prefix pwscf
+      $ mpiexec pyzfs --wfcfmt qeh5 --prefix pwscf
 
    where pwscf is the prefix used for the Quantum Espresso calculation.
 
@@ -66,7 +78,7 @@ After installation, **PyZFS** can be executed in two manners:
 
    .. code:: bash
 
-      $ mpiexec python -m pyzfs.exec.runzfs --wfcfmt qbox --filename gs.xml
+      $ mpiexec pyzfs --wfcfmt qbox --filename gs.xml
 
    where gs.xml is the XML save file generated by Qbox.
 
