### 2020-03-13 5b56c89
add more explanations in the run.sh of examples
