+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name="pyzfs",
-    version="1.0",  # set as dev for development commits
+    version="1.1",
     author="He Ma, Marco Govoni, Giulia Galli",
     author_email="mahe@uchicago.edu, mgovoni@anl.gov, gagalli@uchicago.edu",
     description="A python code to compute zero-field splitting tensors",
