+++ examples/o2_qe_hdf5/run.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 
-# The wavefunction used was generated with QE 6.4.1.
+# The wavefunction used was generated with QE 6.4.1 using the following command
 # pw.x -i pw.in > pw.out
 
+# Run PyZFS to compute the ZFS tensor
 pyzfs --wfcfmt qeh5 > zfs.out
-# equivalently: python -m pyzfs.exec.runzfs --wfcfmt qeh5 > zfs.out
+# An equivalent command is:
+# python -m pyzfs.exec.runzfs --wfcfmt qeh5 > zfs.out
 
 D=`grep --color=never "D unit" zfs.xml | grep --color=never -Eoh '[+-]?[0-9]+([.][0-9]+)?'`
 Dref=`grep --color=never "D unit" zfs_ref.xml | grep --color=never -Eoh '[+-]?[0-9]+([.][0-9]+)?'`
