+++ README.md
@@ -2,5 +2,19 @@
 
 **PyZFS** is a Python package for first-principles calculations of the zero-field-splitting (ZFS) tensor of molecules and materials based on wavefunctions obtained from density functional theory (DFT) calculations. 
 
+Documentation
+------------
+
 The tutorial and documentation are hosted on [Read the Docs](https://pyzfs-doc.readthedocs.io/en/latest/).
 
+Authors
+------------
+He Ma (mahe@uchicago.edu)
+Marco Govoni (mgovoni@anl.gov)
+Giulia Galli (gagalli@uchicago.edu)
+
+Contact, support, and contribution information
+----------------------------------------------
+
+To contact the authors about **PyZFS**, please use the `issue tracker
+<https://github.com/hema-ted/pyzfs/issues>`_ of the GitHub project. Bug reports and contributions to new features are welcome.
