### 2020-03-19 3083324
remove deprecated codes
