+++ pyzfs/common/wfc/qeh5loader.py
@@ -217,89 +217,3 @@ class QEHDF5WavefunctionLoader(WavefunctionLoader):
             pass
         else:
             raise ValueError
-
-    # # deprecated loader (rank 0 send wfc to all other ranks):
-    # def load(self, iorbs, sdm=None):
-    #     super(QEHDF5WavefunctionLoader, self).load(iorbs)
-    #
-    #     # define varibles for MPI communications
-    #     comm = self.comm
-    #     rank = comm.Get_rank()
-    #     size = comm.Get_size()
-    #     onroot = rank == 0
-    #
-    #     iorbs_of_rank = None
-    #     if onroot:
-    #         iorbs_of_rank = {0: iorbs}
-    #         for r in range(1, size):
-    #             iorbs_of_rank[r] = comm.recv(source=r)
-    #     else:
-    #         comm.send(iorbs, dest=0)
-    #
-    #     comm.barrier()
-    #
-    #     # processor 0 parse wavefunctions
-    #     gvecs = ngvecs = psig_arrs_all = None
-    #     if onroot:
-    #         iorbs_all = set.union(*iorbs_of_rank.values())
-    #         psig_arrs_all = {}
-    #
-    #         # read wavefunctions
-    #         c = Counter(len(iorbs_all), percent=0.1,
-    #                     message="(process 0) {n} orbitals ({percent}%) loaded in {dt}...")
-    #         for ispin in range(2):
-    #             wfcfile = "wfcup1.hdf5" if ispin == 0 else "wfcdw1.hdf5"
-    #             wfch5 = h5py.File(os.path.join(
-    #                 self.root, "{}.save".format(self.prefix), wfcfile))
-    #
-    #             gvecs = np.array(wfch5["MillerIndices"], dtype=int)
-    #             ngvecs = gvecs.shape[0]
-    #             assert gvecs.shape == (ngvecs, 3)
-    #
-    #             evc = np.array(wfch5["evc"])
-    #
-    #             for ievc in range(evc.shape[0]):
-    #                 band = ievc + 1
-    #                 iorb = self.wfc.sb_iorb_map.get(("up" if ispin == 0 else "down", band))
-    #
-    #                 if iorb in iorbs_all:
-    #                     psig_arrs_all[iorb] = evc[ievc].view(complex).copy()
-    #                     c.count()
-    #
-    #     # broadcast G vectors
-    #     ngvecs = comm.bcast(ngvecs, root=0)
-    #     if onroot:
-    #         self.wfc.gvecs = gvecs
-    #     if not onroot:
-    #         self.wfc.gvecs = np.zeros([ngvecs, 3], dtype=int)
-    #     comm.Bcast(self.wfc.gvecs, root=0)
-    #
-    #     # scatter wavefunctions
-    #     if rank == 0:
-    #         psig_arrs = {iorb: psig_arrs_all[iorb] for iorb in iorbs}
-    #         for r in range(1, size):
-    #             for iorb in iorbs_of_rank[r]:
-    #                 comm.Send(psig_arrs_all[iorb], dest=r, tag=iorb)
-    #
-    #     else:
-    #         psig_arrs = {iorb: np.zeros(ngvecs, complex) for iorb in iorbs}
-    #         for iorb in iorbs:
-    #             comm.Recv(psig_arrs[iorb], source=0, tag=iorb)
-    #
-    #     comm.barrier()
-    #
-    #     for iorb in iorbs:
-    #         self.wfc.set_psig_arr(iorb, psig_arrs[iorb])
-    #
-    #     if self.memory == "high":
-    #         self.wfc.compute_all_psir()
-    #         self.wfc.clear_all_psig_arr()
-    #         self.wfc.compute_all_rhog()
[…]
