### 2020-03-19 hema-ted/pyzfs#9 @malramsay64
[review] Running unit tests
I would like to see the documentation mention the unit tests you have written and how to run them. Something like

## Tests
There are some unit tests to check functionality which can be run using
```
python -m unittest
```
Additionally the examples directory has examples that check the correct working of this software.
