### 2020-03-19 #2160 @malramsay64
Hi @hema-ted and @dfm,

I have had the opportunity to complete my review. The final (very minor) outstanding issue is https://github.com/hema-ted/pyzfs/issues/9 which will likely be addressed very shortly. Thank you @hema-ted for being so quick to respond and address feedback.
