### 2020-03-19 hema-ted/pyzfs#9 @hema-ted
I've accepted the pull request. Thanks!
