### 2020-03-24 #2160 @xwang862
Hi @hema-ted , thanks for the great work. Both the codes and the summary paper are well structured and clearly written. Most of my comments are trying to increase the accessibility of your package to general users, as well as to satisfy my own curiosity. I will mention them in the order of the checklist:

- General checks:

  - Apart from the `README.md` file, the authors may consider adding a brief description and several topics on the top of the `pyzfs` GitHub repository. It could possibly increase the exposure of your program to a broader audience.

- Functionality:

  - I encountered an error when I was naively following your installation steps. See issue #11
  - Consider adding a `--help` option to the `pyzfs` command.

- Documentation:

  - In the documentation, the authors should add an overview section that talks about the need, the uniqueness of your code among other programs that serve a similar purpose, 1 or 2 application showcase, and how to cite your program (probably after the publication).
  - The tutorial starts with the assumption that users already have certain wavefunction files. It's great that the example files are self-sufficient, but I worry that some users (especially those new to the field) may have trouble generating these files. While such information/tutorial can be found elsewhere, the authors may consider adding an example on how to generate required wfn files. For example, use Quantum Espresso as the solver, show a sample input file of triplet O2, show some `pw.x` commands, etc.
  - In the tutorial, the authors didn't explain `wave` and `density` as supported values of `--fftgrid`. Same goes to `--memory`. Referring users to the source code is usually not a good practice.
  - Can the authors walk through a sample output? E.g. what content does each output file contain, what does each section mean, etc. It doesn't have to be line-by-line, but explaining some core portion of the output will be helpful.

- Software Paper:
  - The authors mention that for systems without heavy elements, i.e. where spin-orbital coupling (SOC) is negligible, the spin-spin interaction (SS) is dominant for the ZFS tensor. But according to [this paper](https://aip.scitation.org/doi/full/10.1063/1.2772857), at least for ground-state O2, the SS and SOC contributions are almost equally important. Can the authors comment on this?
  - The reference(s) of equation 2 should be added.
  - Have the authors compare the accuracy of `pyzfs` to other programs (such as Orca)? How were the reference D values in the example files obtained?
  - Can the authors briefly comment on the performance of `pyzfs` in terms of 1) wall time of a typical calculation, and 2) parallel scaling?
