### 2020-03-25 8914384
use find_packages in setup.py
