### 2020-03-25 36bccf1
explain fftgrid and memory flags in tutorial
