### 2020-03-25 #2160 @hema-ted
Hi @xwang862,

Thanks for your comments and suggestions! I will address some of them here and get back to the rest of them shortly.

help flag: I have included the --help flag (commit e03f6e33).

fftgrid and memory options: I have added more detailed explanations for them in the tutorial (commit 36bccf1b).

wavefunction files: I have included all the input files for DFT codes (QE, Qbox) in my examples. There is a run.sh file in each folder of the /examples directory, which includes the command to perform corresponding DFT calculations to obtain the wavefunction.
