### 2020-03-26 8ea3b3c
update README.md and paper acknowledgement
