### 2020-03-26 c40e919
add a short description in doc for the output file
