+++ doc/tutorial.rst
@@ -83,3 +83,7 @@ After installation, **PyZFS** can be executed in two manners:
    where gs.xml is the XML save file generated by Qbox.
 
 See *pyzfs/examples* for examples of computing the ZFS tensor for the oxygen molecule and the nitrogen-vacancy (NV) center in diamond.
+
+After **PyZFS** is executed, the D tensor, its eigenvalues and eigenvectors are printed by the end of the output.
+The widely-used scalar D and E parameters are also printed.
+A "zfs.xml" file is generated that includes these information, facilitating parsing the results through scripts.
