### 2020-03-26 d8a531d
add ref to eq 2 of paper
