+++ paper/paper.bib
@@ -68,4 +68,9 @@
   year={2019},
   publisher={APS}
 }
-
+@book{Harriman:1978,
+  title={Theoretical foundations of electron spin resonance},
+  author={Harriman, John E},
+  year={1978},
+  publisher={Academic press}
+}
