### 2020-03-26 7379226
add a brief comment on parallelization and timing in tutorial
