### 2020-03-26 #2160 @hema-ted
Hi @xwang862,

Thanks again for your comments. I addressed them in several recent commits. See a summary below.

General checks: I added a , I also expanded the README.md file about the scope of the code.

Documentation - comparison to other codes: Currently, we are not aware of other popular open-sourced code for this type of calculation in the plane-wave basis.

Documentation - output file: I included a short description of the output file by the end of the tutorial.

Paper - O2 and spin-orbital coupling: Indeed, the spin-orbital contribution to O2 is on the same order of magnitude as the spin-spin contribution. It is a general trend that the spin-spin contribution to ZFS tensor dominates the spin-orbit contribution for molecules and materials with light main group elements, and O2 is an (interesting) exception rather than a common case (for example, see https://pubs.acs.org/doi/10.1021/jp0643303 for the calculation of organic molecules, where spin-spin contribution dominates). We used O2 as an example to demonstrate the use of the code because O2 is one of the smallest molecules with a spin-triplet ground state, and the value of its ZFS tensor (both contributions) is well known. Also, for many defects in semiconductors with high point group symmetry, the spin-orbit contributions to the ZFS tensor are zero, and only spin-spin contributions present.

Paper - a reference for eq. 2: I added a reference for eq. 2.

Paper - comparison to other codes like ORCA: PyZFS focuses on computing ZFS tensor using plane-wave basis sets while ORCA uses atomic basis sets. The comparison between ZFS tensors computed using different basis sets is beyond the scope of the current paper, but there are papers (e.g. https://journals.aps.org/prmaterials/abstract/10.1103/PhysRevMaterials.3.043801) that address this issue. The reference values provided in the examples are provided for automatic testing and for verification of installation, and the values are obtained using PyZFS itself.

Paper - timing and parallelization: I added a brief description by the end of the tutorial for timing and parallelization.

Please let me know if you have any further comments.

Finally, I gratefully added the acknowledgment to both reviewers in the paper.
