### 2020-03-26 #2160 @dfm
Actually, the paper looks great to me. So, after you add the DOIs, can you bump the version number, tag a new release on GitHub, and archive that release using Zenodo or similar. Please make sure that the archive *has the same title and author list as the paper* (you might need to edit the metadata if the release is automatically generated). Once you've made these updates, please report the version number and Zenodo DOI here and we'll proceed. Thanks!
