+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={73--78},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wcms.81}
 }
 @article{Rayson:2008,
   title={First principles method for the calculation of zero-field splitting tensors in periodic systems},
@@ -16,7 +17,8 @@
   number={3},
   pages={035119},
   year={2008},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.77.035119}
 }
 @article{Giannozzi:2009,
   title={QUANTUM ESPRESSO: a modular and open-source software project for quantum simulations of materials},
@@ -26,7 +28,8 @@
   number={39},
   pages={395502},
   year={2009},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0953-8984/21/39/395502}
 }
 @article{Gygi:2008,
   title={Architecture of Qbox: A scalable first-principles molecular dynamics code},
@@ -36,7 +39,8 @@
   number={1.2},
   pages={137-144},
   year={2008},
-  publisher={IBM}
+  publisher={IBM},
+  doi={10.1147/rd.521.0137}
 }
 @article{Seo:2017,
   title={Designing defect-based qubit candidates in wide-gap binary semiconductors for solid-state quantum technologies},
@@ -46,7 +50,8 @@
   number={7},
   pages={075002},
   year={2017},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevmaterials.1.075002}
 }
 @article{Whiteley:2019,
   title={Spin-phonon interactions in silicon carbide addressed by Gaussian acoustics},
@@ -56,7 +61,8 @@
   number={5},
   pages={490-495},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41567-019-0420-0}
 }
 @article{Ghosh:2019,
   title={All-electron density functional calculations for electron and nuclear spin interactions in molecules and solids},
@@ -66,7 +72,8 @@
   number={4},
   pages={043801},
   year={2019},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevmaterials.3.043801}
 }
 @book{Harriman:1978,
   title={Theoretical foundations of electron spin resonance},
