### 2020-03-27 #2160 @hema-ted
@dfm I have added the DOIs, changed the version number (1.3) and tagged a new release (named 'joss'), the Zenodo DOI for the release is 10.5281/zenodo.3728346. Please let me know if everything looks right. Thanks!
