### 2020-03-30 #2160 @danielskatz
👋 @hema-ted & @dfm - since it's not so useful to have multiple people proof-reading at the same time, please let us (@openjournals/jose-eics) know when you think this is ready to publish, then we will take a final look and move it forward.
