+++ paper/paper.bib
@@ -79,5 +79,6 @@
   title={Theoretical foundations of electron spin resonance},
   author={Harriman, John E},
   year={1978},
-  publisher={Academic press}
+  publisher={Academic press},
+  doi={https://doi.org/10.1016/c2013-0-06187-7}
 }
