### 2020-03-30 #2160 @dfm
That should be @openjournals/joss-eics! This is ready to go. Thanks!
