### 2020-03-30 #2160 @danielskatz
Please see changes in https://github.com/hema-ted/pyzfs/pull/12 and https://github.com/hema-ted/pyzfs/pull/13
