### 2020-03-30 #2160 @danielskatz
👋 @hema-ted - sorry, I now see a DOI that needs to be fixed, as done in https://github.com/hema-ted/pyzfs/pull/14
